Класс Way.java: * Переименован поле maxSpeed -> speed. * Задан новый метод на реализацию: getEdges(). Его задача из дороги выдать все грани.
Класс Edge.java:
* Изменен тип поля id: int -> long. Соответственно изменен конструктор и методы.

Класс Node.java:
* Изменен тип поля id: int -> long. Соответственно изменен конструктор и методы.

Класс Graph.java:
* Полностью переписаны поля класса. Поле Map<Long, Node> nodeMap - для хранения всех точек, что описывают геометрию дороги. Map<Long, Edge> edgeMap - для хранения всех граней межу узлами в графе. Map<Long, List<Long>> relations хранит id КЛЮЧЕВЫХ узлов в паре с исходящими/входящими гранями. Соответственно изменен конструктор и требуется перекисание методов.

Класс Main.java:
* метод main() воспользовался паркингом и получил дороги из файла.

Класс OsmParser.java:
* Создан метод, который способен определить ключевые точки на дороге.

+++ src/main/java/mapObjects/Edge.java
@@ -4,19 +4,19 @@ import java.util.List;
 
 /**Класс представляет связь между двумя географическими точками (Point).*/
 public class Edge {
-    private final int id;
+    private final long id;
     private final List<Long> idNodes;
-    private final int speed;
+    private final double speed;
     private double length;
 
-    public Edge(int id, List<Long> idNodes, int speed, double length) {
+    public Edge(long id, List<Long> idNodes, int speed, double length) {
         this.id = id;
         this.idNodes = idNodes;
         this.speed = speed;
         this.length = length;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
@@ -28,7 +28,7 @@ public class Edge {
         return idNodes.get(idNodes.size() - 1);
     }
 
-    public int getSpeed(){
+    public double getSpeed(){
         return speed;
     }
